Fix FileSystem.ls on not existing paths

+++ packages/mobile/src/apis/localFile.js
@@ -92,6 +92,9 @@ const deleteAllFiles = async (dir = Dirs.DocumentDir) => {
   for (const dpath of dpaths) {
     const ddpath = deriveFPath(dpath, dir);
 
+    const doExist = await FileSystem.exists(ddpath);
+    if (!doExist) continue;
+
     const fnames = await FileSystem.ls(ddpath);
     for (const fname of fnames) await FileSystem.unlink(ddpath + '/' + fname);
   }
@@ -105,6 +108,9 @@ const listKeys = async (dir = Dirs.DocumentDir) => {
   const keys = [];
   for (const dpath of dpaths) {
     const ddpath = deriveFPath(dpath, dir);
+
+    const doExist = await FileSystem.exists(ddpath);
+    if (!doExist) continue;
 
     const fnames = await FileSystem.ls(ddpath);
     const fpaths = fnames.map(fname => dpath + '/' + fname);
